Create Inner Joins, Where, Group By, and Order By
This SQL query was used to find the top 5 customers in the top 10 cities that paid the highest amount in movie rentals. The top 10 countries and top 10 cities were found first (separate queries) in order to list the top 10 cities in the "where" clause. The result of this query is a table of the top 5 customers that spent the most on movie rentals.